+++ packages/utils/src/lib/git.integration.test.ts
@@ -2,11 +2,19 @@ import { mkdir, rm, writeFile } from 'node:fs/promises';
 import { join } from 'node:path';
 import { type SimpleGit, simpleGit } from 'simple-git';
 import { expect } from 'vitest';
-import { getGitRoot, getLatestCommit, toGitPath } from './git';
+import {
+  getCurrentBranchOrTag,
+  getGitRoot,
+  getLatestCommit,
+  guardAgainstLocalChanges,
+  safeCheckout,
+  toGitPath,
+} from './git';
 import { toUnixPath } from './transform';
 
-describe('git utils', () => {
+describe('git utils in a git repo with a branch and commits', () => {
   const baseDir = join(process.cwd(), 'tmp', 'testing-git-repo');
+  const changesDir = join(baseDir, 'changes-dir');
   let git: SimpleGit;
 
   beforeAll(async () => {
@@ -21,10 +29,24 @@ describe('git utils', () => {
 
     await git.add('README.md');
     await git.commit('Create README');
+
+    await git.checkout(['master']);
   });
 
   afterAll(async () => {
     await rm(baseDir, { recursive: true, force: true });
+  });
+
+  beforeEach(async () => {
+    await git.checkout(['-b', 'feature-branch']);
+    await git.checkout(['master']);
+  });
+
+  afterEach(async () => {
+    // @TODO try why restore/stash/clean/reset hard etc does not work
+    await rm(changesDir, { recursive: true, force: true });
+    await git.checkout(['master']);
+    await git.deleteLocalBranch('feature-branch');
   });
 
   it('should log latest commit', async () => {
@@ -58,6 +80,81 @@ describe('git utils', () => {
   it('should keep relative Unix path as is (already a Git path)', async () => {
     await expect(toGitPath('Backend/API/Startup.cs')).resolves.toBe(
       'Backend/API/Startup.cs',
+    );
+  });
+
+  it('guardAgainstLocalChanges should throw if history is dirty', async () => {
+    await mkdir(changesDir, { recursive: true });
+    await writeFile(join(changesDir, 'change.md'), '# hello-change\n');
+    await expect(guardAgainstLocalChanges(git)).rejects.toThrow(
+      'Working directory needs to be clean before we you can proceed. Commit your local changes or stash them.',
+    );
+  });
+
+  it('guardAgainstLocalChanges should not throw if history is clean', async () => {
+    await expect(guardAgainstLocalChanges(git)).resolves.toBeUndefined();
+  });
+
+  it('safeCheckout should checkout target branch in clean state', async () => {
+    await expect(git.branch()).resolves.toEqual(
+      expect.objectContaining({ current: 'master' }),
+    );
+    await expect(
+      safeCheckout('feature-branch', {}, git),
+    ).resolves.toBeUndefined();
+    await expect(git.branch()).resolves.toEqual(
+      expect.objectContaining({ current: 'feature-branch' }),
+    );
+  });
+
+  it('safeCheckout should throw if history is dirty', async () => {
+    await mkdir(changesDir, { recursive: true });
+    await writeFile(join(changesDir, 'change.md'), '# hello-change\n');
+    await expect(safeCheckout('master', {}, git)).rejects.toThrow(
+      'Working directory needs to be clean before we you can proceed. Commit your local changes or stash them.',
+    );
+  });
+
+  it('safeCheckout should clean local changes and check out to feature-branch', async () => {
+    // needs to get reset to be clean
+    await mkdir(changesDir, { recursive: true });
+    await writeFile(join(changesDir, 'change.md'), '# hello-change\n');
+    // needs to get cleaned to be clean
+    await writeFile(join(baseDir, 'README.md'), '# hello-world-2\n');
+
+    await expect(
+      safeCheckout('feature-branch', { forceCleanStatus: true }, git),
+    ).resolves.toBeUndefined();
+    await expect(git.branch()).resolves.toEqual(
+      expect.objectContaining({ current: 'feature-branch' }),
+    );
+    await expect(git.status()).resolves.toEqual(
+      expect.objectContaining({ files: [] }),
+    );
+  });
+
+  it('getCurrentBranchOrTag should log current branch', async () => {
+    await expect(getCurrentBranchOrTag(git)).resolves.toBe('master');
+  });
+});
+
+describe('git utils in a git repo without a branch and commits', () => {
+  const baseDir = join(process.cwd(), 'tmp', 'testing-git-repo');
+  let git: SimpleGit;
+
+  beforeAll(async () => {
+    await mkdir(baseDir, { recursive: true });
+    git = simpleGit(baseDir);
+    await git.init();
+  });
+
+  afterAll(async () => {
+    await rm(baseDir, { recursive: true, force: true });
+  });
+
+  it('getCurrentBranchOrTag should throw if no branch is given', async () => {
+    await expect(getCurrentBranchOrTag(git)).rejects.toThrow(
+      'Could not get current tag or branch.',
     );
   });
 });