+++ packages/utils/src/lib/reports/constants.ts
@@ -8,3 +8,39 @@ export const SCORE_COLOR_RANGE = {
   YELLOW_MIN: 0.5,
 };
 /* eslint-enable no-magic-numbers */
+
+export const FOOTER_PREFIX = 'Made with ❤ by'; // replace ❤️ with ❤, because of ❤️ has output issues
+export const CODE_PUSHUP_DOMAIN = 'code-pushup.dev';
+export const README_LINK =
+  'https://github.com/flowup/quality-metrics-cli#readme';
+export const reportHeadlineText = 'Code PushUp Report';
+export const reportOverviewTableHeaders = [
+  '🏷 Category',
+  '⭐ Score',
+  '🛡 Audits',
+];
+export const reportRawOverviewTableHeaders = ['Category', 'Score', 'Audits'];
+export const reportMetaTableHeaders: string[] = [
+  'Commit',
+  'Version',
+  'Duration',
+  'Plugins',
+  'Categories',
+  'Audits',
+];
+
+export const pluginMetaTableHeaders: string[] = [
+  'Plugin',
+  'Audits',
+  'Version',
+  'Duration',
+];
+
+// details headers
+
+export const detailsTableHeaders: string[] = [
+  'Severity',
+  'Message',
+  'Source file',
+  'Line(s)',
+];